Update ncrn geospatial directory structure
