+++ Archive/create_ncrn_gis_standard_OLD.py
