+++ Archive/document_gis_directory_OLD.py
