+++ GIS-Library-Documenter/document_gis_directory.py
